+++ task-1/package.json
@@ -13,7 +13,8 @@
     "dependencies": {
         "axios": "^1.7.2",
         "react": "^18.3.1",
-        "react-dom": "^18.3.1"
+        "react-dom": "^18.3.1",
+        "react-router-dom": "^6.24.1"
     },
     "devDependencies": {
         "@types/react": "^18.3.3",
