+++ task-1/src/main.tsx
@@ -1,5 +1,10 @@
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
+import { BrowserRouter } from 'react-router-dom'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
+ReactDOM.createRoot(document.getElementById('root')!).render(
+    <BrowserRouter>
+        <App />
+    </BrowserRouter>,
+)
